Add popup functionality
There's a bit of a disconnect at the moment between my mental model of
the architecture of this codebase and, you know, the architecture
itself. Namely, I can't seem to decide how separate "Core" is from the
rest of the codebase. On top of that, I have a tendency to unnecessarily
separate things that are located wholly within Core from the rest of
Core. I suspect that this will bite me in the future, but maybe I'm
being paranoid.

Let me explain it like this: I can't decide if I should be able to say
`Game.Serv.Ui` or `Game.Ctx.FOO` from within core. Both currently
compile, because that's the way things are set up (Core isn't, like, a
separate .dll or anything). What I want to be able to do is have 100% of
Core work in 100% of cases across all games, or at least all games that
I'm developing simultaneously. But I only ever want to work on one game
at a time, so...

Furthermore, I plan on constantly evolving Core, and on doing so with
the hopes of building a broader and stronger codebase with every
project, so that subsequent projects hit the ground running even faster.

In other words, the more code that I put into Core, the better.

I think that the only thing to do is to not give myself more trouble
than I already have, and to feel free to use constructions like
`Game.Serv` within Core. I suspect/hope that I will be _adding_ services
to Core, not removing them. I won't use `Game.Ctx` within Core, since
session context managers are game specific. I'd love to have a
compiler-based way to do this, but I'll worry about that another time.